Extract cheated syscalls to a separate file. Simplify logic of cheatable syscall handler.
