+++ crates/cheatnet/src/cheatcodes/deploy.rs
@@ -19,7 +19,7 @@ use starknet_api::hash::StarkHash;
 use starknet_api::patricia_key;
 
 use crate::cheatcodes::EnhancedHintError;
-use crate::execution::syscalls::execute_deployment;
+use crate::execution::cheatable_syscall_handler::execute_deployment;
 use starknet_api::core::{ClassHash, ContractAddress};
 use starknet_api::transaction::Calldata;
