+++ crates/cheatnet/src/execution/calls.rs
@@ -16,7 +16,9 @@ use starknet_api::{
     transaction::Calldata,
 };
 
-use super::{entry_point::execute_call_entry_point, syscalls::CheatableSyscallHandler};
+use super::{
+    cheatable_syscall_handler::CheatableSyscallHandler, entry_point::execute_call_entry_point,
+};
 
 // blockifier/src/execution/syscalls/hint_processor.rs:541 (execute_inner_call)
 pub fn execute_inner_call(
