+++ crates/cheatnet/src/execution/mod.rs
@@ -1,10 +1,11 @@
 #[allow(clippy::module_name_repetitions)]
 pub mod cairo1_execution;
 pub mod calls;
+pub mod cheatable_syscall_handler;
+pub mod cheated_syscalls;
 pub mod contract_print;
 pub mod deprecated;
 pub mod entry_point;
 #[allow(clippy::module_name_repetitions)]
 pub mod execution_info;
 pub mod gas;
-pub mod syscalls;
